fix types of exported time and timezone variables

+++ www/fns/client_time_and_timezone.php
@@ -1,6 +1,6 @@
 <?php
 
 function client_time_and_timezone ($user, &$time, &$timezone) {
-    $time = floor(microtime(true) * 1000);
-    $timezone = $user ? $user->timezone : 0;
+    $time = (int)floor(microtime(true) * 1000);
+    $timezone = $user ? (int)$user->timezone : 0;
 }
